requirements.txt
# TuniGuard Dependencies - All Free Tools

# Core Framework
Flask==3.0.0
flask-restful==0.3.10
flask-cors==4.0.0

# Database
SQLAlchemy==2.0.45
Flask-SQLAlchemy==3.1.1

# AI Integration
google-generativeai==0.8.0

# API Documentation
flask-swagger-ui==4.11.1
flasgger==0.9.7.1

# Validation & Security
marshmallow==3.20.1
flask-marshmallow==0.15.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
bcrypt==4.1.1

# HTTP & Utilities
requests==2.31.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
Werkzeug==3.0.1


gunicorn==21.2.0